Hamming Code 구현

+++ BlockInspection/Hamming74.h
@@ -4,8 +4,8 @@
 class Hamming74 : public Block
 {
 protected: /* block */
-	virtual uint64_t input_block() const { return 4; };
-	virtual uint64_t encode_block() const { return 7; };
+	virtual uint8_t input_block() const { return 4; };
+	virtual uint8_t encode_block() const { return 7; };
 
 public: /* virtual */
 	virtual bool encode();
